Move BibliographicLevel and TypeOfRecord in their own module
