+++ lib/bibdata_rs/src/marc/fixed_field/leader/type_of_record.rs
@@ -1,0 +1,69 @@
+use marctk::Record;
+
+pub enum TypeOfRecord {
+    LanguageMaterial,
+    NotatedMusic,
+    ManuscriptNotatedMusic,
+    CartographicMaterial,
+    ManuscriptCartographicMaterial,
+    ProjectedMedium,
+    NonmusicalSoundRecording,
+    MusicalSoundRecording,
+    TwoDimensionalNonProjectableGraphic,
+    ComputerFile,
+    Kit,
+    MixedMaterials,
+    ThreeDimensionalArtifactOrNaturallyOcurringObject,
+    ManuscriptLanguageMaterial,
+}
+
+impl TryFrom<char> for TypeOfRecord {
+    type Error = String;
+
+    fn try_from(value: char) -> Result<Self, Self::Error> {
+        match value {
+            'a' => Ok(Self::LanguageMaterial),
+            'c' => Ok(Self::NotatedMusic),
+            'd' => Ok(Self::ManuscriptNotatedMusic),
+            'e' => Ok(Self::CartographicMaterial),
+            'f' => Ok(Self::ManuscriptCartographicMaterial),
+            'g' => Ok(Self::ProjectedMedium),
+            'i' => Ok(Self::NonmusicalSoundRecording),
+            'j' => Ok(Self::MusicalSoundRecording),
+            'k' => Ok(Self::TwoDimensionalNonProjectableGraphic),
+            'm' => Ok(Self::ComputerFile),
+            'o' => Ok(Self::Kit),
+            'p' => Ok(Self::MixedMaterials),
+            'r' => Ok(Self::ThreeDimensionalArtifactOrNaturallyOcurringObject),
+            't' => Ok(Self::ManuscriptLanguageMaterial),
+            _ => Err(format!("{} is not a valid type of record", value)),
+        }
+    }
+}
+
+impl TryFrom<&Record> for TypeOfRecord {
+    type Error = String;
+    fn try_from(value: &Record) -> Result<Self, Self::Error> {
+        value
+            .leader()
+            .chars()
+            .nth(6)
+            .ok_or("No Type of Record at LDR/06")?
+            .try_into()
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn it_can_identify_manuscript_notated_music() {
+        let manuscript_notated_music =
+            Record::from_breaker("=LDR 02190cdm a2200385 i 4500").unwrap();
+        assert!(matches!(
+            TypeOfRecord::try_from(&manuscript_notated_music),
+            Ok(TypeOfRecord::ManuscriptNotatedMusic)
+        ));
+    }
+}
